[tests] Don't let mono handle aborts.
Intention and result can be far apart 😒

mono tries hard to print helpful information when native crashes occur, but on
Apple platforms we get the much better information from the native crash
report, in particular for AOT'ed code. Mono's helpfulness wouldn't be so bad
if it didn't sometimes hang the app, preventing the OS from creating the
native crash report, thus providing us with less information than if mono
hadn't tried to be helpful in the first place.

So let's just avoid mono altogether for SIGABRT, which will always be a fatal
signal. This way we'll always get a crash report at least.

Ref: https://github.com/mono/mono/issues/15637